+++ src/components/duck-hunt/event-listeners/EventListenerController.js
@@ -1,0 +1,8 @@
+export default class EventListenerController {
+  initEventListeners() {
+    throw new Error('You have to implement the method initEventListeners()!');
+  }
+  removeEventListeners() {
+    throw new Error('You have to implement the method removeEventListeners()!');
+  }
+}